Register CountryRepository in services, move clients config model to Sections folder
