+++ src/Services/Identity/Identity.Api/Configuration/Sections/ClientsConfiguration.cs
@@ -1,4 +1,4 @@
-namespace Identity.Api.Configuration;
+namespace Identity.Api.Configuration.Sections;
 
 public class ClientsConfig
 {
